NMAAS-484 Storage space added to helm install arguments

+++ nmaas-platform/src/main/java/net/geant/nmaas/orchestration/api/model/AppConfigurationView.java
@@ -13,9 +13,9 @@ import lombok.NoArgsConstructor;
 public class AppConfigurationView {
     @JsonProperty("jsonInput")
     private String jsonInput;
-    private Double storageSpace;
+    private Integer storageSpace;
 
-    public void setStorageSpace(Double storageSpace){
+    public void setStorageSpace(Integer storageSpace){
         this.storageSpace = storageSpace;
     }
