+++ nmaas-platform/src/main/java/net/geant/nmaas/orchestration/entities/AppDeploymentSpec.java
@@ -45,7 +45,7 @@ public class AppDeploymentSpec implements Serializable {
     private KubernetesTemplate kubernetesTemplate;
 
     @Column(nullable = false)
-    private Double defaultStorageSpace;
+    private Integer defaultStorageSpace;
 
     private boolean configFileRepositoryRequired;
 }
