+++ nmaas-platform/src/main/java/net/geant/nmaas/portal/api/domain/AppDeploymentSpec.java
@@ -14,7 +14,7 @@ public class AppDeploymentSpec {
 
     private List<AppDeploymentEnv> supportedDeploymentEnvironments;
 
-    private Double defaultStorageSpace;
+    private Integer defaultStorageSpace;
 
     private boolean configFileRepositoryRequired;
